Update kubelet-serving-cert-approver from 0.9.2 to 0.10.0 and enable HA
During the last Talos Linux update I had some issues a certificate
rotation for the kubelet service on one node. The reason was that
the kubelet-serving-cert-approver pod failed due to an OOM event when
all running pods were shifted to the other two nodes. In return, the new
node did not come up because the kubelet certificate was not
automatically approved. By enabeling HA mode, we will have two replicas
available, which should minimize the chance of this to happan again.

While we are at it, let's also upgrade kubelet-serving-cert-approver to
the latest version.

**Note**: Talos does not re-apply extraManifests, so we manually need to
run `kubectl apply -f <new-version-file>` to apply this to our cluster.